% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringFuns.R
\name{scoreBracket}
\alias{scoreBracket}
\title{Score a predicted bracket against an actual
bracket according to some scoring rule}
\usage{
scoreBracket(bracketPredicted, bracketActual, scoringRule = "exponential")
}
\arguments{
\item{bracketPredicted}{a bracket object}

\item{bracketActual}{a bracket object}

\item{scoringRule}{a character string giving scoring rule}
}
\value{
bracket score
}
\description{
Score a predicted bracket against an actual
bracket according to some scoring rule
}
\examples{
score <- scoreBracket( c(1,3,1), c(1,2,1) )
score
}

